git-svn-id: https://svn.origo.ethz.ch/zoodb/trunk@321 b278144e-e56a-4748-98cb-974f122dab01

+++ src/org/zoodb/jdo/internal/DataSerializer.java
@@ -560,7 +560,7 @@ public final class DataSerializer {
     }
 
     private final void serializeOid(Object obj) {
-        _out.writeLong((Long) ((PersistenceCapableImpl)obj).jdoZooGetOid());
+        _out.writeLong(((PersistenceCapableImpl)obj).jdoZooGetOid());
     }
 
     private final void writeClassInfo(Class<?> cls, Object val) {
